+++ parser/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: lgarg <lgarg@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/05 03:03:11 by lgarg             #+#    #+#             */
-/*   Updated: 2021/08/05 03:03:19 by lgarg            ###   ########.fr       */
+/*   Updated: 2021/08/09 15:48:29 by lgarg            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 char	*ft_strtrim(char const *s, const char *set)
 {
 	char	*str;
-	size_t	start;
-	size_t	end;
+	int		start;
+	int		end;
 	int		i;
 
 	if (!s || !set)
